add model inference ingress, base and edit pipeline for model devops.

+++ python/fedml/cli/model_deployment/device_model_deployment.py
@@ -36,45 +36,58 @@ def start_deployment(end_point_id, model_id, model_version,
     # Check whether triton server is running.
     triton_server_is_running = False
     triton_server_container_name = "{}".format(ClientConstants.FEDML_TRITON_SERVER_CONTAINER_NAME_PREFIX)
-    check_triton_server_running_cmds = "{}docker ps |grep {}".format(sudo_prefix, triton_server_container_name)
-    running_process = ClientConstants.exec_console_with_script(check_triton_server_running_cmds,
-                                                               should_capture_stdout=True,
-                                                               should_capture_stderr=True)
-    ret_code, out, err = ClientConstants.get_console_pipe_out_err_results(running_process)
-    if out is not None:
-        out_str = out.decode(encoding="utf-8")
-        if str(out_str) != "":
-            triton_server_is_running = True
+    if not ClientConstants.is_running_on_k8s():
+        check_triton_server_running_cmds = "{}docker ps |grep {}".format(sudo_prefix, triton_server_container_name)
+        running_process = ClientConstants.exec_console_with_script(check_triton_server_running_cmds,
+                                                                   should_capture_stdout=True,
+                                                                   should_capture_stderr=True)
+        ret_code, out, err = ClientConstants.get_console_pipe_out_err_results(running_process)
+        if out is not None:
+            out_str = out.decode(encoding="utf-8")
+            if str(out_str) != "":
+                triton_server_is_running = True
+    else:
+        check_triton_server_running_cmds = "ps -ef |grep {} |wc -l".format(sudo_prefix, triton_server_container_name)
+        running_process = ClientConstants.exec_console_with_script(check_triton_server_running_cmds,
+                                                                   should_capture_stdout=True,
+                                                                   should_capture_stderr=True)
+        ret_code, out, err = ClientConstants.get_console_pipe_out_err_results(running_process)
+        if out is not None:
+            out_str = out.decode(encoding="utf-8")
+            if str(out_str) == "2":
+                triton_server_is_running = True
 
-    if sys_name == "Linux":
-        if not triton_server_is_running:
-            os.system(sudo_prefix + "apt-get update")
-            os.system(sudo_prefix + "apt-get install docker-ce docker-ce-cli containerd.io docker-compose-plugin")
-            os.system("distribution=$(. /etc/os-release;echo $ID$VERSION_ID) \
-          && sudo rm -f /usr/share/keyrings/nvidia-container-toolkit-keyring.gpg;curl -fsSL https://nvidia.github.io/libnvidia-container/gpgkey | sudo gpg --dearmor -o /usr/share/keyrings/nvidia-container-toolkit-keyring.gpg \
-          && curl -s -L https://nvidia.github.io/libnvidia-container/experimental/$distribution/libnvidia-container.list | \
-             sed 's#deb https://#deb [signed-by=/usr/share/keyrings/nvidia-container-toolkit-keyring.gpg] https://#g' | \
-             sudo tee /etc/apt/sources.list.d/nvidia-container-toolkit.list")
-            os.system(sudo_prefix + "apt-get update")
-            os.system(sudo_prefix + "apt-get install -y nvidia-docker2")
-            os.system(sudo_prefix + "systemctl restart docker")
+    # Setup nvidia docker related packages.
+    if not ClientConstants.is_running_on_k8s():
+        if sys_name == "Linux":
+            if not triton_server_is_running:
+                os.system(sudo_prefix + "apt-get update")
+                os.system(sudo_prefix + "apt-get install docker-ce docker-ce-cli containerd.io docker-compose-plugin")
+                os.system("distribution=$(. /etc/os-release;echo $ID$VERSION_ID) \
+              && sudo rm -f /usr/share/keyrings/nvidia-container-toolkit-keyring.gpg;curl -fsSL https://nvidia.github.io/libnvidia-container/gpgkey | sudo gpg --dearmor -o /usr/share/keyrings/nvidia-container-toolkit-keyring.gpg \
+              && curl -s -L https://nvidia.github.io/libnvidia-container/experimental/$distribution/libnvidia-container.list | \
+                 sed 's#deb https://#deb [signed-by=/usr/share/keyrings/nvidia-container-toolkit-keyring.gpg] https://#g' | \
+                 sudo tee /etc/apt/sources.list.d/nvidia-container-toolkit.list")
+                os.system(sudo_prefix + "apt-get update")
+                os.system(sudo_prefix + "apt-get install -y nvidia-docker2")
+                os.system(sudo_prefix + "systemctl restart docker")
 
+    # Convert models from pytorch to onnx format
     convert_model_container_name = "{}_{}_{}".format(ClientConstants.FEDML_CONVERT_MODEL_CONTAINER_NAME_PREFIX,
                                                      str(end_point_id),
                                                      str(model_id))
     running_model_name = ClientConstants.get_running_model_name(end_point_id, model_id,
                                                                 inference_model_name,
                                                                 model_version)
-    convert_model_cmd = "{}docker stop {}; {}docker rm {}; {}docker run --name {} --rm {} -v {}:/project {} " \
-                        "bash -c \"cd /project && convert_model -m /project --name {} " \
-                        "--backend {} --seq-len 16 128 128\"; exit".format(sudo_prefix, convert_model_container_name,
-                                                                           sudo_prefix, convert_model_container_name,
-                                                                           sudo_prefix, convert_model_container_name,
-                                                                           gpu_attach_cmd,
-                                                                           model_storage_local_path,
-                                                                           inference_convertor_image,
-                                                                           running_model_name,
-                                                                           inference_engine)
+    model_storage_local_path = ClientConstants.get_k8s_slave_host_dir(model_storage_local_path)
+    convert_model_cmd = "{}docker stop {}; {}docker rm {}; " \
+                        "{}docker run --name {} --rm {} -v {}:/project " \
+                        "{} bash -c \"cd /project && convert_model -m /project --name {} " \
+                        "--backend {} --seq-len 16 128 128\"; exit". \
+        format(sudo_prefix, convert_model_container_name, sudo_prefix, convert_model_container_name,
+               sudo_prefix, convert_model_container_name, gpu_attach_cmd, model_storage_local_path,
+               inference_convertor_image, running_model_name,
+               inference_engine)
     logging.info("Convert the model to ONNX format: {}".format(convert_model_cmd))
     convert_process = ClientConstants.exec_console_with_script(convert_model_cmd,
                                                                should_capture_stdout=False,
@@ -103,23 +116,30 @@ def start_deployment(end_point_id, model_id, model_version,
 
     # Run triton server
     if not triton_server_is_running:
-        triton_server_cmd = "{}docker stop {}; {}docker rm {}; {}docker run --name {} {} -p{}:8000 " \
-                            "-p{}:8001 -p{}:8002 " \
-                            "--shm-size {} " \
-                            "-v {}:/models {} " \
-                            "bash -c \"pip install transformers && tritonserver " \
-                            "--model-control-mode=poll --repository-poll-secs={} " \
-                            "--model-repository=/models\" ".format(sudo_prefix, triton_server_container_name,
-                                                                   sudo_prefix, triton_server_container_name,
-                                                                   sudo_prefix, triton_server_container_name,
-                                                                   gpu_attach_cmd,
-                                                                   inference_http_port,
-                                                                   inference_grpc_port,
-                                                                   inference_metric_port,
-                                                                   inference_memory_size,
-                                                                   model_serving_dir,
-                                                                   inference_server_image,
-                                                                   ClientConstants.FEDML_MODEL_SERVING_REPO_SCAN_INTERVAL)
+        if ClientConstants.is_running_on_k8s():
+            triton_server_cmd = "pip install transformers && tritonserver " \
+                                "--model-control-mode=poll --repository-poll-secs={} " \
+                                "--model-repository={}". \
+                format(ClientConstants.FEDML_MODEL_SERVING_REPO_SCAN_INTERVAL,
+                       ClientConstants.get_model_serving_dir())
+        else:
+            triton_server_cmd = "{}docker stop {}; {}docker rm {}; {}docker run --name {} {} -p{}:8000 " \
+                                "-p{}:8001 -p{}:8002 " \
+                                "--shm-size {} " \
+                                "-v {}:/models {} " \
+                                "bash -c \"pip install transformers && tritonserver " \
+                                "--model-control-mode=poll --repository-poll-secs={} " \
+                                "--model-repository=/models\" ".format(sudo_prefix, triton_server_container_name,
+                                                                       sudo_prefix, triton_server_container_name,
+                                                                       sudo_prefix, triton_server_container_name,
+                                                                       gpu_attach_cmd,
+                                                                       inference_http_port,
+                                                                       inference_grpc_port,
+                                                                       inference_metric_port,
+                                                                       inference_memory_size,
+                                                                       model_serving_dir,
+                                                                       inference_server_image,
+                                                                       ClientConstants.FEDML_MODEL_SERVING_REPO_SCAN_INTERVAL)
         logging.info("Run triton inference server: {}".format(triton_server_cmd))
         triton_server_process = ClientConstants.exec_console_with_script(triton_server_cmd,
                                                                          should_capture_stdout=False,
@@ -165,8 +185,8 @@ def should_exit_logs(end_point_id, model_id, cmd_type, cmd_process_id, model_nam
             inference_output_url, model_version, model_metadata, model_config = \
                 get_model_info(model_name, inference_engine, inference_port)
             logging.info("Log test for deploying model successfully, inference url: {}, "
-                         "model metadata: {}, model config: {}".format(
-                inference_output_url, model_metadata, model_config))
+                         "model metadata: {}, model config: {}".
+                         format(inference_output_url, model_metadata, model_config))
             if inference_output_url != "":
                 return True
         except Exception as e:
@@ -185,23 +205,24 @@ def log_deployment_result(end_point_id, model_id, cmd_container_name, cmd_type,
     last_out_logs = ""
     last_err_logs = ""
     while True:
-        logs_cmd = "{}docker logs {}".format(sudo_prefix, cmd_container_name)
-        logs_process = ClientConstants.exec_console_with_script(logs_cmd,
-                                                                should_capture_stdout=True,
-                                                                should_capture_stderr=True)
-        ret_code, out, err = ClientConstants.get_console_pipe_out_err_results(logs_process)
-        if out is not None:
-            out_str = out.decode(encoding="utf-8")
-            added_logs = str(out_str).replace(last_out_logs, "")
-            if len(added_logs) > 0:
-                logging.info("{}".format(added_logs))
-            last_out_logs = out_str
-        elif err is not None:
-            err_str = err.decode(encoding="utf-8")
-            added_logs = str(err_str).replace(last_err_logs, "")
-            if len(added_logs) > 0:
-                logging.info("{}".format(added_logs))
-            last_err_logs = err_str
+        if not ClientConstants.is_running_on_k8s():
+            logs_cmd = "{}docker logs {}".format(sudo_prefix, cmd_container_name)
+            logs_process = ClientConstants.exec_console_with_script(logs_cmd,
+                                                                    should_capture_stdout=True,
+                                                                    should_capture_stderr=True)
+            ret_code, out, err = ClientConstants.get_console_pipe_out_err_results(logs_process)
+            if out is not None:
+                out_str = out.decode(encoding="utf-8")
+                added_logs = str(out_str).replace(last_out_logs, "")
+                if len(added_logs) > 0:
+                    logging.info("{}".format(added_logs))
+                last_out_logs = out_str
+            elif err is not None:
+                err_str = err.decode(encoding="utf-8")
+                added_logs = str(err_str).replace(last_err_logs, "")
+                if len(added_logs) > 0:
+                    logging.info("{}".format(added_logs))
+                last_err_logs = err_str
 
         time.sleep(3)
